+++ hyrisecockpit/api/app/__init__.py
@@ -1,6 +1,7 @@
 """Factory method to create an app."""
 
 from flask import Flask
+from flask_cors import CORS
 from flask_restx import Api
 
 
@@ -9,6 +10,7 @@ def create_app(env=None) -> Flask:
     from .routes import register_routes
 
     app: Flask = Flask(__name__)
+    CORS(app)
     api: Api = Api(
         app,
         title="Hyrise Cockpit",
@@ -16,5 +18,5 @@ def create_app(env=None) -> Flask:
         version="2.0",
     )
 
-    register_routes(api, app)
+    register_routes(api, app, root="")
     return app
