+++ hyrisecockpit/api/app/monitor/app.py
@@ -8,17 +8,14 @@ from json import loads
 from time import time_ns
 from typing import Any, Dict, List, Union
 
-from flask import Flask
-from flask_cors import CORS
-from flask_restx import Api, Resource, fields
+from flask_restx import Namespace, Resource, fields
 from influxdb import InfluxDBClient
 from jsonschema import ValidationError, validate
 from zmq import REQ, Context, Socket
 
 from hyrisecockpit.message import get_databases_response_schema, response_schema
-from hyrisecockpit.plugins import available_plugins
 from hyrisecockpit.request import Header, Request
-from hyrisecockpit.response import Response, get_error_response, get_response
+from hyrisecockpit.response import Response, get_response
 from hyrisecockpit.settings import (
     DB_MANAGER_HOST,
     DB_MANAGER_PORT,
@@ -43,22 +40,11 @@ storage_connection = InfluxDBClient(
     STORAGE_HOST, STORAGE_PORT, STORAGE_USER, STORAGE_PASSWORD
 )
 
-app = Flask(__name__)
-cors = CORS(app)
-api = Api(
-    app,
-    title="Hyrise Cockpit",
-    description="Monitor and control multiple databases at once.",
-    validate=True,
-)
-
-monitor = api.namespace(
+api = Namespace(
     "monitor", description="Get synchronous data from multiple databases at once."
 )
 
-control = api.namespace("control", description="Control multiple databases at once.")
-
-model_database = monitor.model(
+model_database = api.model(
     "Database",
     {
         "id": fields.String(
@@ -70,7 +56,7 @@ model_database = monitor.model(
     },
 )
 
-model_throughput = monitor.clone(
+model_throughput = api.clone(
     "Throughput",
     model_database,
     {
@@ -83,13 +69,13 @@ model_throughput = monitor.clone(
     },
 )
 
-model_detailed_throughput = monitor.clone(
+model_detailed_throughput = api.clone(
     "Detailed Throughput",
     model_database,
     {
         "detailed_throughput": fields.List(
             fields.Nested(
-                monitor.model(
+                api.model(
                     "Throughput per query",
                     {
                         "workload_type": fields.String(
@@ -118,13 +104,13 @@ model_detailed_throughput = monitor.clone(
     },
 )
 
-model_query_information = monitor.clone(
+model_query_information = api.clone(
     "Detailed Throughput and Latency",
     model_database,
     {
         "detailed_query_information": fields.List(
             fields.Nested(
-                monitor.model(
+                api.model(
                     "Throughput and latency per query",
                     {
                         "workload_type": fields.String(
@@ -159,7 +145,7 @@ model_query_information = monitor.clone(
     },
 )
 
-model_latency = monitor.clone(
+model_latency = api.clone(
     "Latency",
     model_database,
     {
@@ -172,13 +158,13 @@ model_latency = monitor.clone(
     },
 )
 
-model_detailed_latency = monitor.clone(
+model_detailed_latency = api.clone(
     "Detailed Latency",
     model_database,
     {
         "detailed_latency": fields.List(
             fields.Nested(
-                monitor.model(
+                api.model(
                     "Latency per query",
                     {
                         "workload_type": fields.String(
@@ -207,7 +193,7 @@ model_detailed_latency = monitor.clone(
     },
 )
 
-model_queue_length = monitor.clone(
+model_queue_length = api.clone(
     "Queue length",
     model_database,
     {
@@ -220,7 +206,7 @@ model_queue_length = monitor.clone(
     },
 )
 
-model_workload_composition = monitor.model(
+model_workload_composition = api.model(
     "Workload composition",
     {
         "SELECT": fields.Integer(
@@ -250,7 +236,7 @@ model_workload_composition = monitor.model(
     },
 )
 
-model_krueger_data = monitor.clone(
+model_krueger_data = api.clone(
     "Krüger data",
     model_database,
     {
@@ -269,7 +255,7 @@ model_krueger_data = monitor.clone(
     },
 )
 
-model_database_status = monitor.clone(
+model_database_status = api.clone(
     "Database status",
     model_database,
     {
@@ -295,7 +281,7 @@ model_database_status = monitor.clone(
         ),
         "loaded_tables": fields.List(
             fields.Nested(
-                monitor.model(
+                api.model(
                     "Loaded tables",
                     {
                         "table_name": fields.String(
@@ -318,226 +304,6 @@ model_database_status = monitor.clone(
     },
 )
 
-model_data = control.model(
-    "Data",
-    {
-        "folder_name": fields.String(
-            title="Folder name",
-            description="Name of the folder containing the pregenerated tables.",
-            required=True,
-            example="tpch_0.1",
-        )
-    },
-)
-
-model_storage = control.model(
-    "storage",
-    {
-        fields.String(
-            title="Database ID",
-            description="Used to identify a database.",
-            required=True,
-            example="hyrise-1",
-        ): {
-            fields.String(
-                title="Tablename",
-                description="Name of the table.",
-                required=True,
-                example="aka_name",
-            ): {
-                "size": fields.Integer(
-                    title="Size",
-                    description="Estimated size of the table given in bytes.",
-                    required=True,
-                    example="2931788734",
-                ),
-                "number_columns": fields.Integer(
-                    title="Number of columns",
-                    description="Number of columns of the table.",
-                    required=True,
-                    example="112",
-                ),
-                "data": {
-                    "column_name": {
-                        "size": fields.Integer(
-                            title="Size",
-                            description="Estimated size of the column given in bytes.",
-                            required=True,
-                            example="8593371",
-                        ),
-                        "encoding": fields.String(
-                            title="Encoding",
-                            description="Encodings of the column.",
-                            required=True,
-                            example="Dictionary",
-                        ),
-                        "data_type": fields.String(
-                            title="Datatype",
-                            description="Datatype of the column.",
-                            required=True,
-                            example="String",
-                        ),
-                    }
-                },
-            }
-        }
-    },
-)
-
-model_control_database = control.model(
-    "Database",
-    {
-        "id": fields.String(
-            title="Database ID",
-            description="Used to identify a database.",
-            required=True,
-            example="hyrise-1",
-        )
-    },
-)
-
-model_get_database = control.clone(
-    "Get Database",
-    model_control_database,
-    {
-        "host": fields.String(
-            title="Host",
-            description="Host to log in to.",
-            required=True,
-            example="vm.example.com",
-        ),
-        "port": fields.String(
-            title="Port",
-            description="Port of the host to log in to.",
-            required=True,
-            example="1234",
-        ),
-        "number_workers": fields.Integer(
-            title="Number of initial database worker processes.",
-            description="",
-            required=True,
-            example=8,
-        ),
-        "dbname": fields.String(
-            title="",
-            description="Name of the database to log in to.",
-            required=True,
-            example="mydb",
-        ),
-    },
-)
-
-model_add_database = control.clone(
-    "Add Database",
-    model_get_database,
-    {
-        "user": fields.String(
-            title="Username",
-            description="Username used to log in.",
-            required=True,
-            example="user123",
-        ),
-        "password": fields.String(
-            title="Password",
-            description="Password used to log in.",
-            required=True,
-            example="password123",
-        ),
-    },
-)
-
-modelhelper_plugin = fields.String(
-    title="Plugin name",
-    description="Used to identify a plugin.",
-    required=True,
-    example="Clustering",
-)
-
-model_plugin_log = control.clone(
-    "Plugin Log",
-    model_database,
-    {
-        "plugin_log": fields.List(
-            fields.Nested(
-                control.model(
-                    "Plugin Log Entry",
-                    {
-                        "timestamp": fields.Integer(
-                            title="Timestamp",
-                            description="Timestamp in nanoseconds.",
-                            required=True,
-                            example=1583847966784,
-                        ),
-                        "reporter": fields.String(
-                            title="Reporter",
-                            description="Plugin reporting to the log.",
-                            required=True,
-                            example="CompressionPlugin",
-                        ),
-                        "message": fields.String(
-                            title="Message",
-                            description="Message logged.",
-                            required=True,
-                            example="No optimization possible with given parameters!",
-                        ),
-                    },
-                )
-            ),
-            required=True,
-        )
-    },
-)
-
-model_get_all_plugins = control.model(
-    "Available Plugins", {"plugins": fields.List(modelhelper_plugin, required=True,)},
-)
-
-model_get_activated_plugins = control.clone(
-    "Activated Plugins",
-    model_database,
-    {"plugins": fields.List(modelhelper_plugin, required=True,)},
-)
-
-model_activate_plugin = control.clone(
-    "Activate Plugin", model_database, {"plugin": modelhelper_plugin},
-)
-
-model_deactivate_plugin = control.clone(
-    "Deactivate Plugin", model_database, {"plugin": modelhelper_plugin},
-)
-
-model_plugin_setting = control.clone(
-    "Set Plugin Setting",
-    model_database,
-    {
-        "name": fields.String(
-            title="Setting name",
-            description="Name of the setting that shall be set.",
-            required=True,
-            example="CompressionPlugin_MemoryBudget",
-        ),
-        "value": fields.String(
-            title="Setting value",
-            description="Value the setting should have.",
-            required=True,
-            example="5000",
-        ),
-    },
-)
-
-model_get_plugin_setting = control.clone(
-    "Get Plugin Setting",
-    model_plugin_setting,
-    {
-        "description": fields.String(
-            title="Setting description",
-            description="Description of the plugin setting.",
-            required=True,
-            example="The memory budget to target for the Compression...",
-        ),
-    },
-)
-
 
 def _send_message(socket: Socket, message: Request) -> Response:
     """Send an IPC message with data to a database interface, return the repsonse."""
@@ -556,11 +322,11 @@ def _active_databases() -> List[str]:
     return [database["id"] for database in response["body"]["databases"]]
 
 
-@monitor.route("/throughput")
+@api.route("/throughput")
 class Throughput(Resource):
     """Throughput information of all databases."""
 
-    @monitor.doc(model=[model_throughput])
+    @api.doc(model=[model_throughput])
     def get(self) -> Union[int, Response]:
         """Return throughput information from the stored queries."""
         currentts = time_ns()
@@ -590,11 +356,11 @@ class Throughput(Resource):
         return response
 
 
-@monitor.route("/detailed_throughput")
+@api.route("/detailed_throughput")
 class DetailedThroughput(Resource):
     """Detailed throughput information of all databases."""
 
-    @monitor.doc(model=[model_detailed_throughput])
+    @api.doc(model=[model_detailed_throughput])
     def get(self) -> Union[int, List[Dict[str, Any]]]:
         """Return detailed throughput information from the stored queries."""
         currentts = time_ns()
@@ -623,11 +389,11 @@ class DetailedThroughput(Resource):
         return response
 
 
-@monitor.route("/detailed_latency")
+@api.route("/detailed_latency")
 class DetailedLatency(Resource):
     """Detailed throughput information of all databases."""
 
-    @monitor.doc(model=[model_detailed_latency])
+    @api.doc(model=[model_detailed_latency])
     def get(self) -> Union[int, List[Dict[str, Any]]]:
         """Return detailed throughput information from the stored queries."""
         currentts = time_ns()
@@ -656,11 +422,11 @@ class DetailedLatency(Resource):
         return response
 
 
-@monitor.route("/latency")
+@api.route("/latency")
 class Latency(Resource):
     """Latency information of all databases."""
 
-    @monitor.doc(model=[model_latency])
+    @api.doc(model=[model_latency])
     def get(self) -> Union[int, Response]:
         """Return latency information from the stored queries."""
         currentts = time_ns()
@@ -689,11 +455,11 @@ class Latency(Resource):
         return response
 
 
-@monitor.route("/detailed_query_information")
+@api.route("/detailed_query_information")
 class DetailedQueryInformation(Resource):
     """Detailed throughput and latency information of all databases."""
 
-    @monitor.doc(model=[model_query_information])
+    @api.doc(model=[model_query_information])
     def get(self) -> Union[int, List[Dict[str, Any]]]:
         """Return detailed throughput and latency information from the stored queries."""
         currentts = time_ns()
@@ -724,11 +490,11 @@ class DetailedQueryInformation(Resource):
         return response
 
 
-@monitor.route("/queue_length")
+@api.route("/queue_length")
 class QueueLength(Resource):
     """Queue length information of all databases."""
 
-    @monitor.doc(model=[model_queue_length])
+    @api.doc(model=[model_queue_length])
     def get(self) -> Response:
         """Return queue length information from database manager."""
         return _send_message(
@@ -736,7 +502,7 @@ class QueueLength(Resource):
         )
 
 
-@monitor.route("/failed_tasks")
+@api.route("/failed_tasks")
 class FailedTasks(Resource):
     """Failed tasks information of all databases."""
 
@@ -755,7 +521,7 @@ class FailedTasks(Resource):
         ]
 
 
-@monitor.route("/system")
+@api.route("/system")
 class System(Resource):
     """System data information of all databases."""
 
@@ -784,7 +550,7 @@ class System(Resource):
         return response
 
 
-@monitor.route("/chunks")
+@api.route("/chunks")
 class Chunks(Resource):
     """Chunks data information of all databases."""
 
@@ -810,7 +576,7 @@ class Chunks(Resource):
         return response
 
 
-@monitor.route("/storage")
+@api.route("/storage")
 class Storage(Resource):
     """Storage information of all databases."""
 
@@ -837,11 +603,11 @@ class Storage(Resource):
         return response
 
 
-@monitor.route("/krueger_data", methods=["GET"])
+@api.route("/krueger_data", methods=["GET"])
 class KruegerData(Resource):
     """Krügergraph data for all workloads."""
 
-    @monitor.doc(model=[model_krueger_data])
+    @api.doc(model=[model_krueger_data])
     def get(self) -> Union[int, List[Dict[str, Dict[str, Dict]]]]:
         """Provide mock data for a Krügergraph."""
         krueger_data: List[Dict] = []
@@ -867,227 +633,13 @@ class KruegerData(Resource):
         return krueger_data
 
 
-@monitor.route("/status", methods=["GET"])
+@api.route("/status", methods=["GET"])
 class ProcessTableStatus(Resource):
     """Database blocked status information of all databases."""
 
-    @monitor.doc(model=[model_database_status])
+    @api.doc(model=[model_database_status])
     def get(self) -> List[Dict]:
         """Return status of databases."""
         return _send_message(
             db_manager_socket, Request(header=Header(message="status"), body={}),
         )["body"]["status"]
-
-
-@control.route("/database", methods=["GET", "POST", "DELETE"])
-class Database(Resource):
-    """Manages databases."""
-
-    @control.doc(model=[model_get_database])
-    def get(self) -> Response:
-        """Get all databases."""
-        message = Request(header=Header(message="get databases"), body={})
-        response = _send_message(db_manager_socket, message)
-        return response["body"]["databases"]
-
-    @control.doc(body=model_add_database)
-    def post(self) -> Response:
-        """Add a database."""
-        message = Request(
-            header=Header(message="add database"),
-            body={
-                "number_workers": control.payload["number_workers"],
-                "id": control.payload["id"],
-                "user": control.payload["user"],
-                "password": control.payload["password"],
-                "host": control.payload["host"],
-                "port": control.payload["port"],
-                "dbname": control.payload["dbname"],
-            },
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-    @control.doc(body=model_control_database)
-    def delete(self) -> Response:
-        """Delete a database."""
-        message = Request(
-            header=Header(message="delete database"),
-            body={"id": control.payload["id"]},
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-
-@control.route("/workload", methods=["POST", "DELETE"])
-class Workload(Resource):
-    """Manages workload generation."""
-
-    def post(self) -> Response:
-        """Start the workload generator."""
-        message = Request(header=Header(message="start worker"), body={})
-        response = _send_message(db_manager_socket, message)
-        if response["header"]["status"] != 200:
-            return get_error_response(
-                400, response["body"].get("error", "Error during starting of worker")
-            )
-
-        message = Request(
-            header=Header(message="start workload"),
-            body={
-                "folder_name": control.payload["folder_name"],
-                "frequency": control.payload.get("frequency", 200),
-            },
-        )
-        response = _send_message(generator_socket, message)
-        if response["header"]["status"] != 200:
-            return get_error_response(
-                400,
-                response["body"].get("error", "Error during starting of the workload"),
-            )
-
-        return get_response(200)
-
-    def delete(self) -> Response:
-        """Stop the workload generator and empty database queues."""
-        message = Request(header=Header(message="stop workload"), body={})
-        response = _send_message(generator_socket, message)
-        if response["header"]["status"] != 200:
-            return get_error_response(
-                400, response["body"].get("error", "Error during stopping of generator")
-            )
-
-        message = Request(header=Header(message="close worker"), body={})
-        response = _send_message(db_manager_socket, message)
-        if response["header"]["status"] != 200:
-            return get_error_response(
-                400, response["body"].get("error", "Error during closing of worker")
-            )
-
-        return response
-
-
-@control.route("/data")
-class Data(Resource):
-    """Manage data in databases."""
-
-    @control.doc(model=[model_data])
-    def get(self) -> List[str]:
-        """Return all pregenerated tables that can be loaded."""
-        return ["tpch_0.1", "tpch_1", "tpcds_1", "job"]
-
-    # @control.doc(body=model_data)
-    def post(self) -> Response:
-        """Load pregenerated tables for all databases."""
-        message = Request(
-            header=Header(message="load data"),
-            body={"folder_name": control.payload["folder_name"]},
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-    @control.doc(body=model_data)
-    def delete(self) -> Response:
-        """Delete pregenerated tables from all databases."""
-        message = Request(
-            header=Header(message="delete data"),
-            body={"folder_name": control.payload["folder_name"]},
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-
-@control.route("/available_plugins")
-class ActivatedPlugin(Resource):
-    """Get all available Plugins."""
-
-    @control.doc(model=model_get_all_plugins)
-    def get(self) -> List[str]:
-        """Return available plugins."""
-        return available_plugins
-
-
-@control.route("/plugin")
-class Plugin(Resource):
-    """Activate, Deactive Plugins, respectively show which ones are activated."""
-
-    @control.doc(model=[model_get_activated_plugins])
-    def get(self) -> Union[Dict, List[Dict[str, List[str]]]]:
-        """Return activated plugins in each database."""
-        message = Request(header=Header(message="get plugins"), body={})
-        response = _send_message(db_manager_socket, message)
-        return response["body"]["plugins"]
-
-    @control.doc(body=model_activate_plugin)
-    def post(self) -> Response:
-        """Activate a plugin in a database."""
-        message = Request(
-            header=Header(message="activate plugin"),
-            body={"id": control.payload["id"], "plugin": control.payload["plugin"]},
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-    @control.doc(body=model_deactivate_plugin)
-    def delete(self) -> Response:
-        """Deactivate a plugin in a database."""
-        message = Request(
-            header=Header(message="deactivate plugin"),
-            body={"id": control.payload["id"], "plugin": control.payload["plugin"]},
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
-
-
-@control.route("/plugin_log")
-class PluginLog(Resource):
-    """Activate, Deactive Plugins, respectively show which ones are activated."""
-
-    @api.doc(model=[model_plugin_log])
-    def get(self) -> List[Dict[str, Union[str, List[Dict[str, Union[str, int]]]]]]:
-        """Return activated plugins in each database."""
-        return [
-            {
-                "id": database,
-                "plugin_log": [
-                    {
-                        "timestamp": row["timestamp"],
-                        "reporter": row["reporter"],
-                        "message": row["message"],
-                    }
-                    for row in list(
-                        storage_connection.query(
-                            "SELECT timestamp, reporter, message from plugin_log;",
-                            database=database,
-                        )["plugin_log", None]
-                    )
-                ],
-            }
-            for database in _active_databases()
-        ]
-
-
-@control.route("/plugin_settings")
-class PluginSettings(Resource):
-    """Set settings for plugins."""
-
-    @control.doc(model=[model_get_plugin_setting])  # TODO: fix model
-    def get(self) -> Response:
-        """Read settings for plugins."""
-        message = Request(header=Header(message="get plugin setting"), body={},)
-        response = _send_message(db_manager_socket, message)
-        return response
-
-    @control.doc(body=model_plugin_setting)
-    def post(self) -> Response:
-        """Set settings for plugins."""
-        message = Request(
-            header=Header(message="set plugin setting"),
-            body={
-                "id": control.payload["id"],
-                "name": control.payload["name"],
-                "value": control.payload["value"],
-            },
-        )
-        response = _send_message(db_manager_socket, message)
-        return response
