--- hyrisecockpit/backend/__init__.py
@@ -1,1 +1,0 @@
-"""Module for accessing database monitoring data."""
