+++ setup.py
@@ -6,8 +6,7 @@ setup(
     packages=find_packages(),
     entry_points={
         "console_scripts": [
-            "cockpit-backend=hyrisecockpit.backend.cli:main",
-            "cockpit-api=hyrisecockpit.api.wsgi:main",
+            "cockpit-backend=hyrisecockpit.api.wsgi:main",
             "cockpit-manager=hyrisecockpit.database_manager.cli:main",
             "cockpit-generator=hyrisecockpit.workload_generator.cli:main",
         ],
